PyEasy21: Introduce proper policy
I had forgotten to actually use Q to find the best action to take. It
was all random. Use it now. This also made me realize that I don't need
an explicit representation of the policy. I update Q in every step, so
in every step I have to look into Q for the best action to take. No need
to store that best action in an array.