CljEasy21: Go back to original code, 1M iterations
This is an attempt to get at the real optimum value function by
following the exercise instructions. Also check in the results, because
I don't want to run 1'000'000 iterations too often. The resulting
approximate optimum value function is smoother than before, especially
when the player sum is 10 or below. Further up, it's still rough.

I suspect that this is because player sums ≦ 10 are starting states and
each starting state has probability ≧ 1/100 of being visited at least
once in a game, independent of the policy. The probability of higher-up
sums depends on the policy. I won't make stronger claims about their
probability, since these would be intuitive (I don't want to calculate)
and intuitive claims about probabilities are usually wrong.